% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_address}
\alias{get_address}
\title{Locate an IP adress}
\usage{
get_address(ip, format = "csv")
}
\arguments{
\item{ip}{A single IPv4/IPv6 address or a domain name. If you don't supply a query the current IP address will be used.}

\item{format}{Json, xml, csv, newline or php.}
}
\value{
A response.
}
\description{
Locate an IP adress
}
\examples{
resp <- get_address("142.162.45.64")

resp |>
  httr2::resp_body_string()
}
