% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{get_location}
\alias{get_location}
\title{Get location}
\usage{
get_location(ip, format = "csv")
}
\arguments{
\item{ip}{A single IPv4/IPv6 address or a domain name. If you don't supply a query the current IP address will be used.}

\item{format}{Json, xml, csv, newline or php.}
}
\value{
A response.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Internal function.

For API documentation and terms of service, see \href{https://ip-api.com/}{ip-api.com}.
}
\examples{
resp <- get_location("142.162.45.64")

resp |>
  httr2::resp_body_string()
}
