% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate_ip}
\alias{locate_ip}
\title{Locate an IP adress}
\usage{
locate_ip(ip)
}
\arguments{
\item{ip}{A single IPv4/IPv6 address or a domain name. If you don't supply a query the current IP address will be used.}
}
\value{
A string.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
For API documentation and terms of service, see \href{https://ip-api.com/}{ip-api.com}.
}
\examples{
locate_ip("142.162.45.64")
}
