% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate_ips}
\alias{locate_ips}
\title{Locate multiple IP addresses}
\usage{
locate_ips(
  ips = NULL,
  fields = c("status,message,country,city"),
  lang = "en",
  tidy = TRUE
)
}
\arguments{
\item{fields}{String. Response fields to pass on to the API.}

\item{lang}{String. Response language. An ISO 639 code supported by the API. Defaults to English.}

\item{tidy}{Logical. TRUE to return a tibble. FALSE to return a list.}

\item{ip}{Character vector of either an IPv4 address, IPv6 address.}
}
\value{
A list or tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
For API documentation and terms of service, see \href{https://ip-api.com/}{ip-api.com}.
}
\examples{
locate_ips(c("132.203.167.188", "8.8.8.8", "208.80.152.201")
}
