% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{tidy_location}
\alias{tidy_location}
\title{Tidy location string into a tibble}
\usage{
tidy_location(response = NULL, fields = NULL)
}
\arguments{
\item{response}{Body string response}

\item{fields}{Response fields to pass on to the API.}
}
\value{
Tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Internal function.
}
\examples{
ip <- "142.162.45.64"
fields <- "status,message,country,city"

response <- get_location(ip, fields) |>
   httr2::resp_body_string()

tidy_location(response, fields)

}
