% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_ip.R
\name{validate_ip}
\alias{validate_ip}
\alias{validate_ipv4}
\alias{validate_ipv6}
\title{Validate}
\usage{
validate_ip(ip)

validate_ipv4(ip)

validate_ipv6(ip)
}
\arguments{
\item{ip}{IP address in IPv4 or IPv6}
}
\description{
Validate
}
\section{Functions}{
\itemize{
\item \code{validate_ip()}: Validate IP

\item \code{validate_ipv4()}: Validate IPv4

\item \code{validate_ipv6()}: Validate IPv6

}}
