% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{locate_ip}
\alias{locate_ip}
\title{Locate an IP adress}
\usage{
locate_ip(ip, format = "csv")
}
\arguments{
\item{ip}{A single IPv4/IPv6 address or a domain name. If you don't supply a query the current IP address will be used.}

\item{format}{Json, xml, csv, newline or php.}
}
\value{
A response.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
resp <- locate_ip("142.162.45.64")

resp |>
  httr2::resp_body_string()
}
