% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateip-package.R
\docType{package}
\name{locateip-package}
\alias{locateip}
\alias{locateip-package}
\title{locateip: Locate IP Addresses With ip-api.com}
\description{
Locate IP addresses with ip-api.com.
}
\author{
\strong{Maintainer}: Judith Bourque \email{info@clessn.ca} [copyright holder]

}
\keyword{internal}
