% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateip-package.R
\docType{package}
\name{locateip-package}
\alias{locateip}
\alias{locateip-package}
\title{locateip: Locate IP Addresses With 'ip-api'}
\description{
Download Internet Protocol (IP) address location and more from the 'ip-api' API.
}
\author{
\strong{Maintainer}: Judith Bourque \email{info@clessn.ca} [copyright holder]

}
\keyword{internal}
