% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateip-package.R
\docType{package}
\name{locateip-package}
\alias{locateip}
\alias{locateip-package}
\title{locateip: Locate IP Addresses with 'ip-api'}
\description{
Download Internet Protocol (IP) address location and more from the 'ip-api' application programming interface (API) \url{https://ip-api.com/}. The package makes it easy to get the latitude, longitude, country, region, and organisation associated to the provided IP address. The information is conveniently returned in a rectangular format.
}
\author{
\strong{Maintainer}: Judith Bourque \email{judith.bourque.2@ulaval.ca} (\href{https://orcid.org/0009-0000-3653-1530}{ORCID}) [copyright holder]

}
\keyword{internal}
