% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_ip.R
\name{validate_ip}
\alias{validate_ip}
\alias{validate_ipv4}
\alias{validate_ipv6}
\title{Validate an IP address}
\usage{
validate_ip(ip)

validate_ipv4(ip)

validate_ipv6(ip)
}
\arguments{
\item{ip}{String. IP address as IPv4 or IPv6.}
}
\value{
Logical. TRUE if it's either IPv4 or IPv6, FALSE if not.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Internal function.

Validate an IP adress as IPv4 or IPv6.
}
\section{Functions}{
\itemize{
\item \code{validate_ip()}: Validate IPv4 or IPv6

\item \code{validate_ipv4()}: Validate IPv4

\item \code{validate_ipv6()}: Validate IPv6

}}
